Add VGA PCI driver for QEMU
Couldn't get VGA over ISA bus to work, so this should do instead.
We don't need it for the FPGA, but porting stuff will be much easier
when we can try it out locally first.

This was taken from Mimiker. See the README.md

Now a simple make builds for the Malta board and should display a
yellow Sierpiński fractal at 640x480 resolution