Implemented registers for pipeline
Hacky, but it works.
A painpoint is that there's a spurious write to the register file during
InstructionFetch.
So we need to finetune the time we run the processor for if we are executing
a loop for the testbench, lest the register file we will check in the
testbench reads a wrong value.

I should fix that.