Make RAM accessible to machine code
Stores a byte and loads it and checks it.
Also Registers writes are now "atomic" in regard to the pipeline stages,
so no longer does cutting off the simulation of a loop cause spurious
writes to appear in the register set. (I could probably phrase this better,
but who'll ever read this?)

What's not implemented is reading from ROM in the MEM stage.