Map InstructionMemory on main memory bus
This is configurable over a generic parameter to the CPU.
There's a tiny bug though: The first instruction is omitted when booting.
But just place a NOP there for now and everything is fine ;)
Alternatively, use separate address spaces by settings
SINGLE_ADDRESS_SPACE => false when instantiating the CPU component.