Connect MMIO to memory and add VGA test ROM
ROMs are now under roms/ and selectable via a generic parameter.

The VGA ROM example fails because apparently we have problems with
SB from one register and immediately reading it into another.

Needs to be fixed. Best by checking if it's a general problem
with the main working RAM too.

The MEM_read signal was an unsuccessful attempt to correct it.
Nevertheless, it might be usable in future, so it's kept.