Move VRAM into shader, so it can be accessed asynchronously
There must be a way to do that across modules, but how?

Backstory:
The original idea was having multiple VGA modes.
Each VGA mode consists of a tuple of a Sync mode,
which defines the control signals and thereby the resolution
and a shader, which accesses the VRAM in a specific way
e.g. you can pair a 1024x768 sync with a shader that use a color palette
While this would be cleaner, I think there's no way around having the
shader work at double the VGA frequency, so it may access the VRAM.
As I got no time for that, the mode code is now commented out.